AI: move the guardian spell cast logic from the AI planner to the anonymous namespace
